+++ CHANGELOG.md
@@ -4,6 +4,8 @@ Record of the modification in project development.
 ## [Unreleased] - yyyy-mm-dd
 ### Added
 ### Changed
+- [Issue 26 : Update AN01 for btrfs](https://github.com/suikan4github/kaiten-yaki/issues/26)
+
 ### Deprecated
 ### Removed
 ### Fixed