+++ CHANGELOG.md
@@ -10,6 +10,8 @@ Record of the modification in project development.
 
 ### Deprecated
 ### Removed
+- [Issue 28 : Move application notes to Wiki](https://github.com/suikan4github/kaiten-yaki/issues/28)
+
 ### Fixed
 - [Issue 24 : Fail to install the ubuntu when the / volume is btrfs](https://github.com/suikan4github/kaiten-yaki/issues/24)