+++ CHANGELOG.md
@@ -4,6 +4,14 @@ Record of the modification in project development.
 ## [Unreleased] - yyyy-mm-dd
 ### Added
 ### Changed
+### Deprecated
+### Removed
+### Fixed
+### Known Issue
+
+## [1.2.0] - 2021-10-16
+### Added
+### Changed
 - [Issue 25 : Refactoring: Sourcing config.sys is not needed in the chrooted_job](https://github.com/suikan4github/kaiten-yaki/issues/25)
 - [Issue 26 : Update AN01 for btrfs](https://github.com/suikan4github/kaiten-yaki/issues/26)
 - [Issue 27 : Eliminates the confirmation dialog](https://github.com/suikan4github/kaiten-yaki/issues/27)
@@ -66,6 +74,7 @@ See [Testing before release v1.1.0](https://github.com/suikan4github/kaiten-yaki
 ### Known Issue
 
 
-[Unreleased]: https://github.com/suikan4github/kaiten-yaki/compare/v1.1.0...develop
+[Unreleased]: https://github.com/suikan4github/kaiten-yaki/compare/v1.2.0...develop
+[1.2.0]: https://github.com/suikan4github/kaiten-yaki/compare/v1.1.0...v1.2.0
 [1.1.0]: https://github.com/suikan4github/kaiten-yaki/compare/v1.0.0...v1.1.0
 [1.0.0]: https://github.com/suikan4github/kaiten-yaki/compare/v0.0.0...v1.0.0